backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg2-binary
pydantic
httpx
google-generativeai>=0.8.0
google-genai
